this version is good but long

+++ libft/libft.h
@@ -69,7 +69,7 @@ t_list				*ft_lstnew(void *content);
 t_list				*ft_lstmap(t_list *lst, void *(*f)(void *),
 											void (*del)(void *));
 int					ft_isspace(int c);
-char				*ft_utoa_base(unsigned int n, char *base);
+char				*ft_utoa_base(unsigned long n, char *base);
 
 
 #endif
